+++ packages/Bio/src/const.ts
@@ -1,5 +1,3 @@
 import * as grok from 'datagrok-api/grok';
 import * as ui from 'datagrok-api/ui';
 import * as DG from 'datagrok-api/dg';
-
-export const GAP_SYMBOL: string = '';